requirements.txt
Django
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
requests
djangorestframework